20 Questions Game
This program implements MVC design and XML file reading and writing.
The rules of this game are: the program asks the user a series of
questions to guess the name of a restaurant that the user has in mind.
If the restaurant is unknown, the program adds it to its database of
restaurants